+++ src/main/java/rs/hybridit/controller/BookCopyController.java
@@ -4,6 +4,7 @@ package rs.hybridit.controller;
 import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -30,40 +31,52 @@ public class BookCopyController {
 	}
 
 	@PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE)
-	@ResponseStatus(HttpStatus.CREATED)
-	public BookCopy create(@RequestBody BookCopyDto bookCopyDto) {
-		return bookCopyService.create(new BookCopy(bookCopyDto));
+	public ResponseEntity<?> create(@RequestBody BookCopyDto bookCopyDto) {
+		BookCopy bookCopy = new BookCopy(bookCopyDto);
+		return new ResponseEntity<>(bookCopy, HttpStatus.CREATED);
 	}
 
 	@GetMapping(value = "/{id}")
-	@ResponseStatus(HttpStatus.OK)
-	public BookCopy getBookCopy(@PathVariable Long id) {
-		BookCopy bookCopy = bookCopyService.getOne(id);
-		return bookCopy;
+	public ResponseEntity<?> getBookCopy(@PathVariable Long id) {
+		BookCopy bookCopy = bookCopyService.findById(id);
+		if (bookCopy != null) {
+			return new ResponseEntity<>(bookCopy, HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+		}
 	}
 
 	@GetMapping
-	@ResponseStatus(HttpStatus.OK)
-	public List<BookCopy> getBookCopys() {
-		return bookCopyService.getAll();
+	public ResponseEntity<?> getBookCopys() {
+		List<BookCopy> bookCopies = bookCopyService.getAll();
+		if (bookCopies != null) {
+			return new ResponseEntity<>(bookCopies, HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+		}
 	}
 
 	@PutMapping(value = "/{id}", consumes = MediaType.APPLICATION_JSON_VALUE)
-	@ResponseStatus(HttpStatus.OK)
-	public BookCopy updateBookCopy(@PathVariable Long id, @RequestBody BookCopyDto bookCopyDto) {
-		BookCopy bookCopy = bookCopyService.getOne(id);
+	public ResponseEntity<?> updateBookCopy(@PathVariable Long id, @RequestBody BookCopyDto bookCopyDto) {
+		BookCopy bookCopy = bookCopyService.findById(id);
 		if (bookCopy != null) {
 			bookCopy.setRentStart(bookCopyDto.getRentStart());
 			bookCopy.setRentEnd(bookCopyDto.getRentEnd());
+			return new ResponseEntity<>(bookCopy, HttpStatus.OK);
 		}
-		return bookCopy;
+		return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
 	}
 
 	@DeleteMapping(value = "/{id}")
-	@ResponseStatus(HttpStatus.OK)
-	public BookCopy deleteBookCopy(@PathVariable Long id) {
-		BookCopy bookCopy = bookCopyService.getOne(id);
-		bookCopyService.delete(bookCopy);
-		return bookCopy;
+	public ResponseEntity<?> deleteBookCopy(@PathVariable Long id) {
+		BookCopy bookCopy = bookCopyService.findById(id);
+		if (bookCopy != null) {
+			bookCopyService.delete(bookCopy);
+			return new ResponseEntity<>(bookCopy, HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+		}
+
 	}
+
 }
